fixed energies
